refactor: form에서 중복되는 스타일을 커스텀 훅으로 분리
