+++ src/styles/store/StoreStyle.ts
